+++ admin-ui/src/app/modules/employees/components/salary/salary-list/salary-list.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
 import { IEmployeeSalary } from '../../../model/employeeModels';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-salary-list',
@@ -16,7 +17,12 @@ export class SalaryListComponent implements OnInit {
   public employeeSalaryColumns: string[] = ['id', 'name', 'designation', 'status', 'salary'];
   public employeeSalaryDataSource: MatTableDataSource<IEmployeeSalary>;
   public employeeSalaries: IEmployeeSalary[] = [];
-  constructor(private http: HttpClient, private router: Router) { }
+
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit() {
     this.getEmployeeSalaries();
@@ -60,18 +66,29 @@ export class SalaryListComponent implements OnInit {
   }
 
   public onDeleteEmployee(employeeSalaryId: number) {
+    const confirmationData: IConfirmation = {
+      title: 'Delete Salary payment',
+      subtitle: 'Are you really sure to delete this salary payment?'
+    };
 
-    this.http.delete(ApiEndpoint.EMPLOYEE_SALARY + '/' + employeeSalaryId).subscribe(data => {
-      this.employeeSalaries = [];
-      this.employeeSalaryDataSource = new MatTableDataSource([]);
-      this.getEmployeeSalaries();
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.EMPLOYEE_SALARY + '/' + employeeSalaryId).subscribe(data => {
+            this.employeeSalaries = [];
+            this.employeeSalaryDataSource = new MatTableDataSource([]);
+            this.getEmployeeSalaries();
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 }
