+++ admin-ui/src/app/modules/income-expense/components/income/income-list/income-list.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { Router } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { IIncome } from '../../../models/income-expense.model';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-income-list',
@@ -16,7 +17,12 @@ export class IncomeListComponent implements OnInit {
   public incomeColumns: string[] = ['id', 'refNo', 'amount', 'incomeType', 'incomeDetails', 'comments', 'incomeDate', 'action'];
   public incomeDataSource: MatTableDataSource<IIncome>;
   public incomes: IIncome[] = [];
-  constructor(private http: HttpClient, private router: Router) { }
+
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit() {
     this.getIncomes();
@@ -61,17 +67,29 @@ export class IncomeListComponent implements OnInit {
 
   public onDeleteRow(incomeId: number) {
 
-    this.http.delete(ApiEndpoint.INCOMES + '/' + incomeId).subscribe(data => {
-      this.incomeDataSource = new MatTableDataSource([]);
-      this.getIncomes();
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Income',
+      subtitle: 'Are you really sure to delete this income?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.INCOMES + '/' + incomeId).subscribe(data => {
+            this.incomeDataSource = new MatTableDataSource([]);
+            this.getIncomes();
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 
 }
