+++ admin-ui/src/app/modules/income-expense/components/income/income-view/income-view.component.ts
@@ -2,7 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router, ActivatedRoute } from '@angular/router';
 import { IIncome } from '../../../models/income-expense.model';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { MatDialog } from '@angular/material';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-income-view',
@@ -15,7 +17,12 @@ export class IncomeViewComponent implements OnInit {
   incomeId: string;
   income: IIncome;
 
-  constructor(private http: HttpClient, private router: Router, private activatedRoute: ActivatedRoute) {
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog
+  ) {
     this.activatedRoute.params.subscribe(params => {
       this.incomeId = params.id;
     });
@@ -43,16 +50,28 @@ export class IncomeViewComponent implements OnInit {
 
   delete() {
 
-    this.http.delete(ApiEndpoint.INCOMES + '/' + this.incomeId).subscribe(data => {
-      this.router.navigate(['/admin/incomes-expenses/incomes']);
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Income',
+      subtitle: 'Are you really sure to delete this income?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.INCOMES + '/' + this.incomeId).subscribe(data => {
+            this.router.navigate(['/admin/incomes-expenses/incomes']);
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 
   edit() {
