+++ admin-ui/src/app/modules/maintenances/admission-fee-structure/admissionfee-structure-list/admissionfee-structure-list.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { Router } from '@angular/router';
 import { IAdmissionFeeMaintenanceYearly } from 'src/app/modules/students/models/maintenance';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-admissionfee-structure-list',
@@ -16,7 +17,12 @@ export class AdmissionfeeStructureListComponent implements OnInit {
   public admFeeStructureColumns: string[] = ['id', 'year', 'action'];
   public admFeeStructureDataSource: MatTableDataSource<IAdmissionFeeMaintenanceYearly>;
   public admFeeStructures: IAdmissionFeeMaintenanceYearly[] = [];
-  constructor(private http: HttpClient, private router: Router) { }
+
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit() {
     this.getAdmFeeStructures();
@@ -59,16 +65,28 @@ export class AdmissionfeeStructureListComponent implements OnInit {
 
   public onDeleteRow(admFeeStructureId: number) {
 
-    this.http.delete(ApiEndpoint.ADDMISSION_FEE_MAINTENANCE + '/' + admFeeStructureId).subscribe(data => {
-      this.admFeeStructureDataSource = new MatTableDataSource([]);
-      this.getAdmFeeStructures();
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Admission Fee Structure',
+      subtitle: 'Are you really sure to delete this admission fee structure?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.ADDMISSION_FEE_MAINTENANCE + '/' + admFeeStructureId).subscribe(data => {
+            this.admFeeStructureDataSource = new MatTableDataSource([]);
+            this.getAdmFeeStructures();
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 }
