+++ admin-ui/src/app/modules/maintenances/admission-fee-structure/admissionfee-structure-view/admissionfee-structure-view.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router, ActivatedRoute } from '@angular/router';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { IAdmissionFeeMaintenanceYearly, IAdmissionFeeMaintenance } from 'src/app/modules/students/models/maintenance';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-admissionfee-structure-view',
@@ -20,7 +21,12 @@ export class AdmissionfeeStructureViewComponent implements OnInit {
   public admFeeStructureDataSource: MatTableDataSource<IAdmissionFeeMaintenance>;
   public admFeeStructures: IAdmissionFeeMaintenance[] = [];
 
-  constructor(private http: HttpClient, private router: Router, private activatedRoute: ActivatedRoute) {
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog
+  ) {
     this.activatedRoute.params.subscribe(params => {
       this.admFeeStructureId = params.id;
     });
@@ -51,16 +57,28 @@ export class AdmissionfeeStructureViewComponent implements OnInit {
 
   delete() {
 
-    this.http.delete(ApiEndpoint.ADDMISSION_FEE_MAINTENANCE + '/' + this.admFeeStructureId).subscribe(data => {
-      this.router.navigate(['/admin/maintenances/admission-fees']);
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Admission Fee Structure',
+      subtitle: 'Are you really sure to delete this admission fee structure?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.ADDMISSION_FEE_MAINTENANCE + '/' + this.admFeeStructureId).subscribe(data => {
+            this.router.navigate(['/admin/maintenances/admission-fees']);
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 
   edit() {
