+++ admin-ui/src/app/modules/students/components/admission-fee/admission-fee-list/admission-fee-list.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
 import { IAdmission } from '../../../models/admission-fee.model';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-admission-fee-list',
@@ -16,7 +17,12 @@ export class AdmissionFeeListComponent implements OnInit {
   public admissionColumns: string[] = ['id', 'name', 'registrationNo', 'Class', 'Admission Number', 'action'];
   public admissionsDataSource: MatTableDataSource<IAdmission>;
   public admissions: IAdmission[] = [];
-  constructor(private http: HttpClient, private router: Router) { }
+
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit() {
     this.getAdmissions();
@@ -61,17 +67,29 @@ export class AdmissionFeeListComponent implements OnInit {
 
   public onDeleteRow(admId: number) {
 
-    this.http.delete(ApiEndpoint.ADMISSIONS + '/' + admId).subscribe(data => {
-      this.admissionsDataSource = new MatTableDataSource([]);
-      this.ngOnInit();
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Admission',
+      subtitle: 'Are you really sure to delete this admission?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.ADMISSIONS + '/' + admId).subscribe(data => {
+            this.admissionsDataSource = new MatTableDataSource([]);
+            this.ngOnInit();
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
