+++ admin-ui/src/app/modules/students/components/admission-fee/admission-fee-view/admission-fee-view.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
-import { MatTableDataSource } from '@angular/material';
+import { MatTableDataSource, MatDialog } from '@angular/material';
 import { IAdmission, IFee } from '../../../models/admission-fee.model';
-import { ApiEndpoint } from 'src/app/modules/shared/model/shared.model';
+import { ApiEndpoint, IConfirmation } from 'src/app/modules/shared/model/shared.model';
 import { HttpClient } from '@angular/common/http';
+import { ConfirmationDialogComponent } from 'src/app/modules/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-admission-fee-view',
@@ -31,7 +32,12 @@ export class AdmissionFeeViewComponent implements OnInit {
   public feesDataSource: MatTableDataSource<IFee>;
   public fee: IFee[] = [];
 
-  constructor(private http: HttpClient, private router: Router, private activatedRoute: ActivatedRoute) {
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog
+  ) {
     this.activatedRoute.params.subscribe(params => {
       this.admId = params.id;
     });
@@ -64,16 +70,28 @@ export class AdmissionFeeViewComponent implements OnInit {
 
   delete() {
 
-    this.http.delete(ApiEndpoint.ADMISSIONS + '/' + this.admId).subscribe(data => {
-      this.router.navigate(['/admin/students/admissions']);
-    }, err => {
-      console.error(err);
-      if (err.error && err.error.apiMessage) {
-        this.errorMessage = err.error.apiMessage.detail;
-      } else {
-        this.errorMessage = err.message;
-      }
-    });
+    const confirmationData: IConfirmation = {
+      title: 'Delete Admission',
+      subtitle: 'Are you really sure to delete this admission?'
+    };
+
+    this.dialog.open(ConfirmationDialogComponent, { width: '26%', data: confirmationData, disableClose: true })
+      .afterClosed().subscribe(okData => {
+        if (okData) {
+
+          this.http.delete(ApiEndpoint.ADMISSIONS + '/' + this.admId).subscribe(data => {
+            this.router.navigate(['/admin/students/admissions']);
+          }, err => {
+            console.error(err);
+            if (err.error && err.error.apiMessage) {
+              this.errorMessage = err.error.apiMessage.detail;
+            } else {
+              this.errorMessage = err.message;
+            }
+          });
+
+        }
+      });
   }
 
   edit() {
